This diff introduce mapView as type representing map objects in simple functions interface.
mapView has similar limited interface as std::map<key, std::optional<val>> .

- begin()
- end()
- size()
- and element *it can be accessed using first and second.
- elements can be compared with std::pair<key, std::optional<value>>.

This diff includes some refactoring:
- move view types to a different header file.
- OptionalVectorValueAccessor: represents values in mapVies and elements in array.
 (a lazy optional like access to vector elements), comparable and assignable to std:optional and has
 similar interface as std::optional.
 - IndexBasedIterator: unify iterator implementation for both mapView and arrayView.

## Performance:
Went down from 20X slower to 2X-3X, PR #584 address major part of that and shows that the slowdown is not 
from the mapView implementation but other parts in the vectorAdapter. 

a new benchmark nestedMapSumVectorFunctionMapView is added:
which uses a vectorReader inside a vector function (used to isolate potential overhead of vector adapter).
PR shows interesting difference between nestedMapSumSimpleFunction and nestedMapSumVectorFunctionMapView

Before:
```
============================================================================
mapSumVectorFunction                                               2.72ms   368.27
mapSumSimpleFunction                                     4.44%    61.16ms    16.35
============================================================================
```
After:
```
============================================================================
mapSumVectorFunction                                         2.97ms   337.17
mapSumSimpleFunction                              40.09%     7.40ms   135.17
nestedMapSumVectorFunction                                   6.19ms   161.47
nestedMapSumSimpleFunction                        26.78%    23.12ms    43.25
nestedMapSumVectorFunctionMapView                 29.53%    20.97ms    47.68
============================================================================
```