@laithsakka 2021-11-08
this was reviewed by @funrollloops  as part of PR#584 
https://github.com/facebookincubator/velox/pull/584